Add Invaders Expansion. Correct for anti-ban bias
Adds support for factions and player mats for Invaders from Afar (Togawa/Albion and Innovative/Militant).
Corrects for mild bug that biased against expansions that could have a banned faction. Previously it would re-roll factions and player mats if there was a banned combo. But now it selects all factions first, and then player mats. If any player mats result in a banned combo it will re-roll the player mats to match with the factoins.